requirements.txt
numpy
scipy
scikit-learn
numpydoc
matplotlib
setuptools
sphinx
sphinx_rtd_theme
sphinx_gallery
pillow
attrs
